✨ feat: add FileList and FileUpload components for report file management
- Introduce FileList component to display and manage files associated with reports, including download and delete functionalities.
- Implement FileUpload component to facilitate file uploads for reports with drag-and-drop support.
- Update ReportsTab to integrate FileList and FileUpload components for enhanced file management experience.
